[Encoding] Allow `PadEncodingAttribute` to support dynamic padding. (#20662)
The dynamic padding allows different backends to chose different padding
values.

To be able to use `?` for representing dynamic values, the container is
changed to be int64_t.

+++ compiler/src/iree/compiler/Codegen/ExternalInterfaces/GPUEncodingExternalModels.cpp
@@ -539,7 +539,7 @@ struct GPUPadEncodingLayoutResolverAttrInterface final
            "Incorrect pad amount");
     assert(padBytes < cacheSetSpanBytes && "Incorrect pad amount");
     const int64_t numPadElements = (padBytes * 8) / elementBits;
-    SmallVector<int32_t> padValues(rank, 0);
+    SmallVector<int64_t> padValues(rank, 0);
     padValues[padDimensionIndex] = numPadElements;
     auto padLayout = Encoding::PadEncodingLayoutAttr::get(ctx, padValues);
     return padLayout;
